+++ main.h
@@ -1,0 +1,7 @@
+#ifndef MAIN_H
+#define MAIN_H
+
+#include "gpmanager.h"
+#include "mainwidget.h"
+
+#endif // MAIN_H
